fixes error of deleting files that were already organized

+++ src/outfit_generator.py
@@ -3,22 +3,25 @@ import os
 import pygame
 
 
-def sort_items(folder):
+def organize_items(folder):
     folder_path = os.path.abspath(folder)
     img_number = 1
 	
     for filename in os.listdir(folder_path):
         img_path = os.path.join(folder_path, filename)
 		
-        with Image.open(img_path) as img:
-            converted_img = img.convert("RGB")
+        if filename.startswith(folder) and filename.endswith(".jpg"):
+            img_number += 1
+        else:
+            with Image.open(img_path) as img:
+                converted_img = img.convert("RGB")
 
-            new_filename = f"{folder}_{img_number}.jpg"
-            new_path = os.path.join(folder_path, new_filename)
+                new_filename = f"{folder}_{img_number}.jpg"
+                new_path = os.path.join(folder_path, new_filename)
 
-            converted_img.save(new_path, "JPEG")
+                converted_img.save(new_path, "JPEG")
             os.remove(img_path)
-        img_number += 1
+            img_number += 1
 
 
 def run_program(screen, bg_tile):
@@ -38,7 +41,9 @@ def main():
 	resolution = (1920, 1080)
 	screen = pygame.display.set_mode(resolution)
 	bg_tile = (pygame.image.load("background/background.png").convert())
-	sort_items('tops')
+	organize_items('tops')
+	organize_items('bottoms')
+	organize_items('shoes')
 	run_program(screen, bg_tile)
